fixes the Grid energy meter issue
Bug: grid meter keeps increasing when No Grid, or Auto modes are selected.

Updates the stopMeter() function. It now takes a bool argument. If the argument is true, then we will increase the meter before stopping it, otherwise, we will not increase the meter, and stop it.

RelayController now checks if the gridRelay is turned on before stopping the meter. If it is, then we add the last interval to the meter, and stop the meter, otherwise, we just stop the meter, without adding the last meter interval.