adds error handling for 85C temp readings
We now hide the 85C erroneous temperature reading due to interference.

To still identify a problem we introduced a new variable that counts the occurrence of 85C, and after 10 consecutive 85C readings we will raise the NO_BOILER_TEMP error and set the temperature to 0C, thus producing the 4608 error code.