adds lost temperature sensor safety
adds a safety feature to turn off nectarsun when the temperature sensor signal is lost while in operation. if the temperature sensor resets and is present again, the error is cleared. also, this feature does not impact the normal operation of the device, if the temperature does not change due to the large volume of the boiler