increases the lost sensor count
increases the LOST_SENSOR_COUNT const to 100, which translates to 300s, or 5mins. if during the 5mins period the temperature does not change between every measurement, nectarsun will ping the temperature sensor to check if it's present. if it is present nectarsun will reset the counter, if it's not present it will set appropriate errors and stop operation.